# Faster collection of NTuples