First of all, this is my first PR to julia, so please be patient with my if I should have missed something.

This PR addresses a topic that came up in #49671.

Currently, dynamic conversion of NTuples to Vectors by `collect()` triggers a new compilation
This PR offers a julia solution to this problem.

There might be other solutions to this issue, most probably compiler-related, but that's certainly out of reach for me.

Before the patch `collect(t::Tuple)` is recompiled for each number of tuple elements
```julia
julia> n = 0;

julia> n += 1; @time collect(tuple(string.(1:n)...),);
  0.110986 seconds (79.05 k allocations: 5.343 MiB, 99.61% compilation time)

julia> n += 1; @time collect(tuple(string.(1:n)...),);
  0.012787 seconds (241 allocations: 11.375 KiB, 99.69% compilation time)
```
after the patch, compilation happens only once:
```julia    
julia> n += 1; @time collect(tuple(string.(1:n)...),);
  0.029778 seconds (18.11 k allocations: 1.207 MiB, 99.68% compilation time)

julia> n += 1; @time collect(tuple(string.(1:n)...),);
  0.000037 seconds (15 allocations: 688 bytes)
```