@JeffBezanson 2023-08-17
Welcome, thank you for looking into this. We do want to avoid excessive compilation time, but I worry that this will be slower for small tuples. Maybe we want a size cutoff so we stop specializing for tuples beyond a certain length?