@hhaensel 2023-08-18
Thanks for your comments, @JeffBezanson.

I digged a little deeper on performance (which I should have done before posting this...). But then I would have missed your comment and the following learning
- `t isa NTuple` is costly and increases with the tuple length
The performance drops significantly, up to factors of 10 and more. So I'm not sure whether switching the method for higher number of elements is that reasonable.
(Would it be possible to have the compiler track whether a Tuple is an NTuple...?)

Then I thought, we might offer only a specialisation of `collect(String, tuple)`, but this would again need ` isa NTuple`, as `collect(String, itr)` already exists.

So if ever, I would tend to offer a separate function `collect_ntuple(T, ntuple)`. However, this function is still a factor 4-5 slower than the specialised collect(T, itr). The developer would need to decide, whether a largely varying number of tuple elements is expected.

Finally I wondered, whether I could somehow find a faster solution and ended up in defining an `unsafe_collect_ntuple(T, ntuple)`. That one is indeed faster, so an experienced developer might benefit from such a function.

(I need to acknowledge the author of this post on stack overflow https://stackoverflow.com/questions/54889057/how-to-get-a-ptr-to-an-element-of-an-ntuple/54898127#54898127)
```julia
function unsafe_ntuple_copyto!(dest::Array{T}, doffs, @nospecialize(src::Tuple), soffs, n) where T
    n == 0 && return dest
    t1 = Base.@_gc_preserve_begin dest
    t2 = Base.@_gc_preserve_begin src

    destp = pointer(dest, doffs)
    srcp = ccall(:jl_value_ptr, Ptr{Cvoid}, (Any,), src)
    soffs == 1 || (srcp += n * Base.elsize(T[]))
    if !Base.allocatedinline(T)
        ccall(:jl_array_ptr_copy, Cvoid, (Any, Ptr{Cvoid}, Any, Ptr{Cvoid}, Int),
              dest, destp, src, srcp, n)
    elseif isbitstype(T)
        ccall(:memmove, Ptr{Cvoid}, (Ptr{Cvoid}, Ptr{Cvoid}, Csize_t),
              destp, srcp, n * Base.aligned_sizeof(T))
    elseif isbitsunion(T)
        ccall(:memmove, Ptr{Cvoid}, (Ptr{Cvoid}, Ptr{Cvoid}, Csize_t),
              destp, srcp, n * Base.aligned_sizeof(T))
        # copy selector bytes
        ccall(:memmove, Ptr{Cvoid}, (Ptr{Cvoid}, Ptr{Cvoid}, Csize_t),
              ccall(:jl_array_typetagdata, Ptr{UInt8}, (Any,), dest) + doffs - 1,
              ccall(:jl_array_typetagdata, Ptr{UInt8}, (Any,), src) + soffs - 1,
              n)
    else
        _unsafe_copyto!(dest, doffs, src, soffs, n)
    end
    Base.@_gc_preserve_end t2
    Base.@_gc_preserve_end t1
    return dest
end

function collect_ntuple(::Type{T}, @nospecialize(t::Tuple)) where T
    res = Vector{T}(undef, length(t))
    i = 1
    for x in t
        @inbounds res[i] = x::T
        i += 1
    end
    res
end

function unsafe_collect_ntuple(::Type{T}, @nospecialize(t::Tuple)) where T
    return unsafe_ntuple_copyto!(Vector{T}(undef, length(t)), 1, t, 1, length(t))
end

function collect_tuple(@nospecialize t::Tuple)
    if isempty(t)
        []
    elseif t isa NTuple
        unsafe_collect_ntuple(typeof(t[1]), t)
    else
        Base._collect(1:1 #= Array =#, t, Base.HasEltype(), Base.HasLength())
    end
end
```


```julia
julia> n = 10000
julia> t1 = tuple(string.(1:n)...);
julia> t2 = tuple(:a, string.(1:n-1)...);

julia> @btime collect(t1);
  13.600 μs (2 allocations: 78.17 KiB)

julia> @btime collect_tuple(t1);
  235.600 μs (4 allocations: 78.20 KiB)

julia> @btime collect(String, t1);
  14.500 μs (2 allocations: 78.17 KiB)

julia> @btime collect_ntuple(String, t1);
  51.800 μs (2 allocations: 78.17 KiB)

julia> @btime unsafe_collect_ntuple(String, t1);
  7.733 μs (3 allocations: 78.19 KiB)

# test whether unsafe_collect_ntuple will be recompiled for a new number of elements
julia> n += 1; t1 = tuple(string.(1:n)...); @time unsafe_collect_ntuple(String, t1);
  0.000024 seconds (3 allocations: 78.250 KiB)
```
So my current proposal would be to define new functions `collect_ntuples()` and `unsafe_collect_ntuples()` and keep the latter unexported. (`collect_tuples()` was just defined here to show the bad performance of `t is NTuple` if the function is not specialised.)

Perhaps there are other ways to speed up `collect_ntuple()`?