@vtjnash 2023-08-18
Adding unsafe to a name doesn't give a function license to be that much wrong. I can see which function you were trying to replicate (from Arrays), but that implementation has a lot of problems when used with Tuple. For starters, `jl_value_ptr` is not legal to use directly since it returns invalid pointers when called on an immutable object. That stack overflow answer is wrong to suggest that. The Ref trick is slightly better, but still has likely many problems when applied in that way.