@hhaensel 2023-08-18
Thanks for the feedback. I did try the ref trick, but that increased compile time significantly and, moreover, reintroduced recompilation.
I did test various scenarios and did not observe any failures. But that's, of course, no proof.
Would you then agree to define `collect_ntuple()` or is it too special and people should take care themselves?
I wonder, whether there might be any other possibility of speeding up collection of NTuples. But that's probably followed up in #49671.

Anyhow, I did take my learnings from your feedback, so thanks again.